requirements.txt
asgiref==3.6.0
Django==3.2.16
django-environ==0.9.0
gunicorn==20.1.0
pytz==2022.7.1
sqlparse==0.4.3
django-filter==21.1
djangorestframework-simplejwt==5.2.2
PyMySQL==1.0.2
djangorestframework==3.12.4
django-cors-headers
mysqlclient
waitress==2.1.0
python-decouple==3.8
django-allauth==0.60.0
django-rest-auth==0.9.5
Pillow
python-dotenv==0.19.0
django-cors-headers